Equipment, Def stats, adjusted stats of the player and enemy.
Equipment is now purchasable and equipable, equipment limited by
on-level-up stat. The def stat was fully implemented so  that players
and and enemies alike have the potential for damage mitigation
(currently only in the form of a +1def equipment). Also reduced player
and enemy damage values, along side enemy health and reinforcements.
Player melee and ranged attacks now do the same amount of damage, with
different multipliers on their skill variations.